Remove PoolingSubmitter.Conditions.heartIsBeating
This implementation looks at the initial heartbeat set during cluster
creation. Hearbeat on the cluster is not updated and should not be used
for such a check. A proper check would look at the master node for each
cluster.

For now, we'll accept the small window where we could select a cluster
while it has already decided to self-destruct but it has not yet updated
it's state to `DELETING` yet. This transition usually is visible within
seconds from submitting a deletion.