Use deterministic names for pooling clusters
This PR changes the way new clusters are created and jobs are assigned
to them in the poolable clusters in order to fix a concurrency problem
that made it create more clusters than the defined limit. The new
implementation uses deterministic names based of the current time to
decide the names of the clusters that should be used at a given moment
in time. Existing clusters than don't belong to them will be deleted
automatically after the max-idle time has passed, as no more jobs will
be assigned to them.